accounts/serializers.py
"""Account serializers"""
from rest_framework import serializers
from django.contrib.auth import get_user_model
from common.serializers import AuditableSerializer

User = get_user_model()

class UserRegistrationSerializer(serializers.ModelSerializer):
    """User registration"""
    password = serializers.CharField(write_only=True, min_length=8)
    password_confirm = serializers.CharField(write_only=True, min_length=8)

    class Meta:
        model = User
        fields = ["email", "username", "password", "password_confirm"]

    def validate(self, attrs):
        if attrs["password"] != attrs.pop("password_confirm"):
            raise serializers.ValidationError({"password": "Passwords don't match"})
        return attrs

    def create(self, validated_data):
        return User.objects.create_user(**validated_data)

class UserDetailSerializer(AuditableSerializer):
    """User details"""
    class Meta:
        model = User
        fields = ["id", "username", "email", "is_verified", "created_at"]